refactor to simplify
