+++ src/renderers/render-aem.js
@@ -1,0 +1,52 @@
+import jsdom from 'jsdom';
+import {
+  isAbsoluteURL,
+  relativeToAbsolute,
+} from '../modules/utils/link-utils.js';
+import isBinary from '../modules/utils/media-utils.js';
+
+/**
+ * Renders content from AEM UE pages
+ */
+export default async function renderAem(path, params, config) {
+  const { authorization } = params;
+  const { aemEnv, owner, repo, ref } = config;
+
+  const aemURL = `${aemEnv}/bin/franklin.delivery/${owner}/${repo}/${ref}${path}?wcmmode=disabled`;
+  const url = new URL(aemURL);
+
+  const fetchHeaders = { 'cache-control': 'no-cache' };
+  if (authorization) {
+    fetchHeaders.authorization = authorization;
+  }
+
+  const resp = await fetch(url, { headers: fetchHeaders });
+
+  if (!resp.ok) {
+    return { error: { code: resp.status, message: resp.statusText } };
+  }
+
+  let contentType = resp.headers.get('content-type') || 'text/html';
+  [contentType] = contentType.split(';');
+
+  const respHeaders = {
+    'content-type': contentType,
+  };
+
+  if (isBinary(contentType)) {
+    const data = Buffer.from(await resp.arrayBuffer());
+    return { data, respHeaders };
+  }
+
+  const html = await resp.text();
+
+  // FIXME: Converting images from AEM to absolue path. Revert once product fix in place.
+  const dom = new jsdom.JSDOM(html);
+  const { document } = dom.window;
+  const elements = document.querySelectorAll('img');
+  elements.forEach((el) => {
+    const uri = el.getAttribute('src');
+    if (!isAbsoluteURL(uri)) el.src = relativeToAbsolute(uri, aemEnv);
+  });
+  return { html: dom.serialize() };
+}
